requirements.txt
numpy==1.21.0
scikit-learn==0.24.2
nltk==3.6.2
pyyaml==5.4.1
spacy==3.0.6
fuzzywuzzy==0.18.0
pyinstaller==4.3
xlrd==2.0.1
pandas==1.2.5
flask==2.0.1
python-Levenshtein==0.12.2
flasgger==0.9.4
transformers==4.8.2
tabulate==0.8.9
sentence-transformers==1.2.0
statsmodels==0.12.2
jupyterlab==3.0.16
matplotlib==3.4.2
pytest==6.2.4
scipy==1.7.0
seaborn==0.11.2
pyinstrument==3.4.2
